added curves to visualization

+++ nagi/neat.py
@@ -301,7 +301,7 @@ class Population(object):
         self.genomes = new_population_of_genomes
         self.speciate()
 
-    def assign_number_of_offspring_to_species(self, fitnesses: Dict[int, float]) -> Dict[int, float]:
+    def assign_number_of_offspring_to_species(self, fitnesses: Dict[int, float]) -> Dict[int, int]:
         total_adjusted_fitness = self._get_total_sum_of_adjusted_fitnesses(fitnesses)
         sum_of_adjusted_fitnesses_by_species = self._get_sum_of_adjusted_fitnesses_by_species(fitnesses)
         assigned_number_of_offspring = {
